migrate fsm tests
