+++ backend/controller/dal/testdata/go/fsmretry/go.mod
@@ -42,5 +42,5 @@ require (
 	google.golang.org/protobuf v1.34.1 // indirect
 )
 
-replace github.com/TBD54566975/ftl => ../../../..
+replace github.com/TBD54566975/ftl => ../../../../../..
